@@ -61,7 +61,8 @@
   },
   "dependencies": {
     "lodash": "^4.17.4",
-    "react": "^15.3.2",
+    "prop-types": "^15.5.10",
+    "react": "^15.5.0",
     "react-dom": "^15.3.2"
   }
 }
